new commit

+++ src/BmiService.java
@@ -1,5 +1,5 @@
 public class BmiService {
-    public double calculate (int weight, int height) {
+    public double calculate (int weight, double height) {
 
         double index = weight / (height * height);
